reorganize xyk tests
