+++ pallets/xyk/src/lib.rs
@@ -40,13 +40,7 @@ use orml_traits::{MultiCurrency, MultiCurrencyExtended};
 use primitives::Amount;
 
 #[cfg(test)]
-mod mock;
-
-#[cfg(test)]
 mod tests;
-
-#[cfg(test)]
-mod invariants;
 
 mod benchmarking;
