+++ pallets/xyk/src/tests/mod.rs
@@ -1,0 +1,3 @@
+mod invariants;
+pub(crate) mod mock;
+mod tests;
